Add Pimoroni and motor2040 libraries to CMakeLists
In order to support Pimoroni's Pico SDK and the motor2040 library, the CMakeLists.txt file was updated to include these libraries in the build. The PIMORONI_PICO_PATH environment variable was added to locate the necessary Pimoroni files. motor2040 was added as a mandatory library for the osod_motor_2040 executable. Also, the SDK got initialized right after the CMake minimum version is confirmed. Lastly, CPack configuration has been added on the bottom of the CMakeLists.txt file to generate compressed files which includes binary files and README for release packages.