Add Receiver library with CPPM decoder functionality
The Receiver library with CPPM decoder functionality was added. This library includes receiver's header and CPPM decoder related functionalities. Also, it includes a method to get receiver data. A third-party library pico_cppm is used for the decoding. The CMakeLists.txt file was also added to handle the dependencies. This library will enable receiving and decoding CPPM signals.