Refactor receiver code to support multiple protocols
The receiver code was refactored to support multiple protocols. Specifically, support was added for both the CPPM and SBUS receiver protocols.  This was achieved by creating a Receiver interface and implementing two receiver classes, "ReceiverCPPM" and "ReceiverSBUS", which inherit from it. This will allow additional receiver protocols to be easily supported in the future, by implementing additional Receiver classes.

A new library, "sbus_2040", was also created to handle SBUS functionality. The existing cppm decoder library is used for cppm functionality. Now, the receiver protocol is determined by a build flag (RX_PROTOCOL) in the CMake file.

The "Navigator" class was also modified to accept a pointer to a receiver object, rather than a constant reference, allowing it to use the new Receiver interface.

In addition, new functions were added to map receiver values to appropriate ranges. This will standardize the range of values for different receiver protocols, allowing them to be used interchangeably in the rest of the codebase.

These changes are necessary to make the codebase more adaptable to different hardware configurations, as different radio transmitters/receivers may use different protocols.