First step in adding StateManager to manage navigated state
StateManager has been added to manage and request the navigated state from the receiver. This change is necessary to further abstract the handling of the incoming receiver data into a separate module. Navigator now also links to statemanager and forwards receiver data to it.

Note that the state manager does not actually do any state management yet, and nor does it send commands to the wheels, as that work is still to be done.