Add RX protocol switch in CMakeLists.txt
A new option has been added to the receiver library's CMakeLists.txt to enable the choice between SBUS and CPPM protocols. This option is supplied as a cache variable RX_PROTOCOL. The default value if none is provided is CPPM. If SBUS is chosen, a new compile definition -DRX_PROTOCOL_SBUS is added. For the default CPPM, the definition is -DRX_PROTOCOL_CPPM. This addition was necessary to provide flexibility in using different receiver protocols.